requirements.txt
anyio==4.11.0
certifi==2025.11.12
et_xmlfile==2.0.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
openpyxl==3.1.5
python-telegram-bot==22.5
sniffio==1.3.1
typing_extensions==4.15.0
xlrd==1.2.0
